Fix lint issues
Removed dependency cycle.
Removed unused imports.
Added missing properties to object initialization.
