+++ src/service/payment.ts
@@ -6,7 +6,7 @@ import {
 } from '@/model/order';
 import {
   BankwirePayIn,
-  Card, CardDirectPayIn, CardDirectPayInCommand, CardRegistration, CardRegistrationCommand, PayIn,
+  Card, CardDirectPayIn, CardDirectPayInCommand, CardRegistration, CardRegistrationCommand,
 } from '@/model/payment';
 
 /**
@@ -18,9 +18,9 @@ export default class PaymentApi extends Api {
   }
 
   /**
-   * Create a new order from the content of the authenticated user's cart. 
+   * Create a new order from the content of the authenticated user's cart.
    * The method requires user to have  role ROLE_CONSUMER.
-   * 
+   *
    * Invoking checkout method multiple times will always return a new order. Only the last created order
    * is valid.
    */
@@ -84,7 +84,7 @@ export default class PaymentApi extends Api {
    * It is imperative to inform users, that we are registering their card.
    *
    * @param command Card registration command
-   * @returns 
+   * @returns
    */
   public async completeCardRegistration(command: CardRegistrationCommand): Promise<ServerResponse<Card>> {
     const url = '/action/payments/cards';
@@ -115,14 +115,14 @@ export default class PaymentApi extends Api {
    *
    * For details on MANGOPAY card direct PayIn see:
    * https://docs.mangopay.com/endpoints/v2.01/payins#e278_create-a-card-direct-payin
-   * 
+   *
    * For details on testing card direct payments see:
    * https://docs.mangopay.com/guide/3ds2-overview
    * https://docs.mangopay.com/guide/3ds2-testing-in-sandbox
    *
    * @param orderKey A unique order key generated by a checkout operation
    * @param command Card direct PayIn command
-   * @returns 
+   * @returns
    */
   public async createCardDirectPayIn(orderKey: string, command: CardDirectPayInCommand): Promise<ServerResponse<CardDirectPayIn>> {
     const url = `/payments/card-direct/${orderKey}`;
@@ -139,12 +139,12 @@ export default class PaymentApi extends Api {
    * Create a bankwire PayIn for a specific order
    * On successful execution, the server resets the user cart; Hence, the client
    * must either reset the cart locally or invoke [Get Cart](#operation/cart-01).
-   * 
+   *
    * For details on MANGOPAY bankwire PayIn see:
    * https://docs.mangopay.com/endpoints/v2.01/payins#e280_create-a-bankwire-direct-payin
-   * 
+   *
    * @param orderKey A unique order key generated by a checkout operation
-   * @returns 
+   * @returns
    */
   public async createBankwirePayIn(orderKey: string): Promise<ServerResponse<BankwirePayIn>> {
     const url = `/payments/bankwire/${orderKey}`;
@@ -159,9 +159,9 @@ export default class PaymentApi extends Api {
 
   /**
    * Get PayIn details
-   * 
+   *
    * @param payInKey A unique PayIn key
-   * @returns 
+   * @returns
    */
   public async getPayIn(payInKey: string): Promise<ServerResponse<BankwirePayIn | CardDirectPayIn>> {
     const url = `/payments/payins/${payInKey}`;
